fix according to code review
