+++ generators/add-action/campaign-standard/index.js
@@ -27,7 +27,7 @@ class CampaignStandardGenerator extends ActionGenerator {
 
     this.addAction(this.props.actionName, './getAllProfiles.js', {
       testFile: './getAllProfiles.test.js',
-      e2eTestFile: './getAllProfiles.e2etest.js',
+      e2eTestFile: './getAllProfiles.e2e.js',
       tplContext: this.props,
       dotenvStub: {
         label: 'please provide your Adobe I/O Campaign Standard integration tenant',
