+++ generators/app/templates/package.json
@@ -7,7 +7,6 @@
       "jest": "^24.9.0"
   },
   "scripts": {
-      "test": "jest",
-      "e2e": "jest --collectCoverage=false --testRegex './e2e/e2e.js'"
+      "test": "jest"
   }
 }
