+++ lib/ActionGenerator.js
@@ -153,7 +153,7 @@ class ActionGenerator extends Generator {
 
   _writeActionE2ETest (actionName, tplActionTest, actionDestPath, tplContext = {}) {
     // test/actions/actionName.test.js
-    const testDestPath = this.destinationPath('e2e', actionsDirname, actionName + '.e2etest.js')
+    const testDestPath = this.destinationPath('e2e', actionsDirname, actionName + '.e2e.js')
     // enriches tplContext with actionRelPath to be required from test file
     tplContext = { actionRelPath: path.relative(path.dirname(testDestPath), actionDestPath), ...tplContext }
     // should we support other options of copyTpl (templateOptions && copyOptions) ?
