+++ test/lib/ActionGenerator.test.js
@@ -345,7 +345,7 @@ describe('implementation', () => {
       actionGenerator.addAction('myAction', './templateFile.js', { e2eTestFile: './template.test.js' })
 
       // test manifest update with action information
-      expect(actionGenerator.fs.copyTpl).toHaveBeenCalledWith(n('/fakeTplDir/template.test.js'), n(`/fakeDestRoot/e2e/${constants.actionsDirname}/myAction.e2etest.js`), { actionRelPath: n(`../../${constants.actionsDirname}/myAction/index.js`) }, {}, {})
+      expect(actionGenerator.fs.copyTpl).toHaveBeenCalledWith(n('/fakeTplDir/template.test.js'), n(`/fakeDestRoot/e2e/${constants.actionsDirname}/myAction.e2e.js`), { actionRelPath: n(`../../${constants.actionsDirname}/myAction/index.js`) }, {}, {})
     })
 
     test('with testFile option and tplContext option (should append relative path to tested file to test template context)', () => {
